Refactor Go host: add TclObj/TclInterp types, eliminate global registry
This commit restructures the Go host implementation for better type safety
and cleaner architecture.

Key changes:

1. Added explicit TclObj and TclInterp types (interp/tclc.go:36-40)
   - Both are Handle (uintptr) underneath, mirroring C's TclHandle
   - Provides type safety when passing handles around

2. Changed TclHandle from uint32_t to uintptr_t (src/tclc.h:90)
   - Required for cgo.Handle which needs pointer-sized values
   - Allows direct use of Go's cgo.Handle for interpreter lookup

3. Eliminated global interpreter registry
   - Was: global map[Handle]*Interp with mutex
   - Now: cgo.Handle stores *Interp directly, lookup is O(1)
   - getInterp() is now trivial: cgo.Handle(h).Value().(*Interp)
   - Added Interp.Close() to release cgo.Handle when done

4. Moved Host type from interp/default/ to interp/host.go
   - Host is core infrastructure, not milestone-specific
   - interp/default/default.go now just returns preconfigured M1 host

5. Unified CommandFunc signature (interp/tclc.go:48)
   - func(i *Interp, cmd TclObj, args []TclObj) TclResult
   - UnknownHandler is now just CommandFunc type
   - Commands set result via i.SetResultString() and return TclResult
   - No more error return - errors are communicated via ResultError

6. Added TclResult as proper Go type with constants
   - ResultOK, ResultError, ResultReturn, ResultBreak, ResultContinue

7. Added Interp methods for result handling:
   - SetResult(obj TclObj)
   - SetResultString(s string)
   - SetErrorString(s string)
   - GetString(h TclObj) string
   - InternString(s string) TclObj

The dispatch flow is now cleaner:
  C interpreter -> goBindUnknown -> UnknownHandler (CommandFunc)
  -> dispatch looks up command -> calls CommandFunc
  -> command sets result via Interp methods -> returns TclResult

TODO for colleague:
- Convert args TclObj (list) to []TclObj slice in goBindUnknown
- This requires implementing list operations in the host

+++ src/tclc.h
@@ -87,7 +87,7 @@
  */
 
 /** An opaque handle type, used by the host to identify objects */
-typedef uint32_t TclHandle;
+typedef uintptr_t TclHandle;
 
 /** A handle to an interpreter instance */
 typedef TclHandle TclInterp;
